+++ lib/lono/cfn/base.rb
@@ -4,14 +4,13 @@ class Lono::Cfn::Base
   include Lono::Cfn::AwsService
   include Lono::Cfn::Util
 
-  attr_reader :randomize_stack_name
   def initialize(stack_name, options={})
-    @randomize_stack_name = options[:randomize_stack_name]
-    @stack_name = randomize(stack_name)
-    @options = options
+    @options = options # options must be set first because @option used in append_suffix
+    stack_name = switch_current(stack_name)
+    @stack_name = append_suffix(stack_name)
     Lono::ProjectChecker.check unless options[:lono] # already ran checker in lono generate
 
-    @template_name = options[:template] || derandomize(@stack_name)
+    @template_name = options[:template] || remove_suffix(@stack_name)
     @param_name = options[:param] || @template_name
     @template_path = get_source_path(@template_name, :template)
     @param_path = get_source_path(@param_name, :param)
@@ -19,7 +18,18 @@ class Lono::Cfn::Base
     puts "Using parameters: #{@param_path}" unless @options[:mute_using]
   end
 
+  def switch_current(stack_name)
+    Lono::Cfn::Current.name!(stack_name)
+  end
+
+  def starting_message
+    action = self.class.to_s.split('::').last
+    action = action[0..-2] + 'ing' # create => creating
+    puts "#{action} #{@stack_name.colorize(:green)} stack..."
+  end
+
   def run
+    starting_message
     params = generate_all
     begin
       save_stack(params) # defined in the sub class
@@ -35,6 +45,13 @@ class Lono::Cfn::Base
         exit 1
       end
     end
+
+    return unless @options[:wait]
+    status.wait
+  end
+
+  def status
+    @status ||= Lono::Cfn::Status.new(@stack_name)
   end
 
   def prompt_for_iam(capabilities)
@@ -55,6 +72,7 @@ class Lono::Cfn::Base
       generate_templates
       unless @options[:noop]
         upload_scripts
+        upload_files
         upload_templates
       end
     end
@@ -80,6 +98,11 @@ class Lono::Cfn::Base
   def upload_scripts
     return unless s3_folder
     Lono::Script::Upload.new.run
+  end
+
+  def upload_files
+    return unless s3_folder
+    Lono::FileUploader.new.upload_all
   end
 
   def generate_params(options={})
@@ -172,28 +195,61 @@ class Lono::Cfn::Base
     exit signal
   end
 
-  # Do nothing unless in Create class
-  def randomize(stack_name)
-    stack_name
+  # Appends a short suffix at the end of a stack name.
+  # Lono internally strips this same suffix for the template name.
+  # Makes it convenient for the development flow.
+  #
+  #   lono cfn current --suffix 1
+  #   lono cfn create demo => demo-1
+  #   lono cfn update demo-1
+  #
+  # Instead of typing:
+  #
+  #   lono cfn create demo-1 --template demo
+  #   lono cfn update demo-1 --template demo
+  #
+  # The suffix can be specified at the CLI but can also be saved as a
+  # preference.
+  #
+  # A random suffix can be specified with random. Example:
+  #
+  #   lono cfn current --suffix random
+  #   lono cfn create demo => demo-[RANDOM], example: demo-abc
+  #   lono cfn update demo-abc
+  #
+  # It is not a default setting because it might confuse new lono users.
+  def append_suffix(stack_name)
+    suffix = Lono.suffix == 'random' ? random_suffix : Lono.suffix
+    [stack_name, suffix].compact.join('-')
   end
 
-  # Strip the random string at end of the template name
-  def derandomize(template_name)
-    if randomize_stack_name?
-      template_name.sub(/-(\w{3})$/,'') # strip the random part at the end
+  def remove_suffix(stack_name)
+    return stack_name unless Lono.suffix
+
+    if stack_name_suffix == 'random'
+      stack_name.sub(/-(\w{3})$/,'') # strip the random suffix at the end
+    elsif stack_name_suffix
+      pattern = Regexp.new("-#{stack_name_suffix}$",'')
+      stack_name.sub(pattern, '') # strip suffix
     else
-      template_name
+      stack_name
     end
   end
 
-  def randomize_stack_name?
-    if !randomize_stack_name.nil?
-      return randomize_stack_name # CLI option takes highest precedence
+  # only generate random suffix for Lono::Cfn::Create class
+  def random_suffix
+    return nil unless self.class.name.to_s =~ /Create/
+    (0...3).map { (65 + rand(26)).chr }.join.downcase # Ex: jhx
+  end
+
+  def stack_name_suffix
+    if @options[:suffix] && !@options[:suffix].nil?
+      return @options[:suffix] # CLI option takes highest precedence
     end
 
     # otherwise use the settings preference
     settings = Lono::Setting.new
-    settings.data['randomize_stack_name']
+    settings.data['stack_name_suffix']
   end
 
   def capabilities
@@ -214,5 +270,27 @@ class Lono::Cfn::Base
   def s3_folder
     setting = Lono::Setting.new
     setting.s3_folder
+  end
+
+  # Either set the templmate_body or template_url attribute based on
+  # if template was uploaded to s3.
+  # Nice to reference s3 url because the max size of the template body is
+  # greater if the template body is on s3. Limits:
+  #
+  #   template_body: 51,200 bytes
+  #   template_url: 460,800 bytes
+  #
+  # Reference: https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/cloudformation-limits.html
+  def set_template_body!(params)
+    # if s3_folder is set this means s3 upload is enabled
+    if s3_folder # s3_folder defined in cfn/base.rb
+      upload = Lono::Template::Upload.new
+      url = upload.s3_presigned_url(@template_path)
+      params[:template_url] = url
+    else
+      params[:template_body] = IO.read(@template_path)
+    end
+
+    params
   end
 end