+++ lib/lono/cfn/delete.rb
@@ -3,7 +3,7 @@ class Lono::Cfn::Delete
   include Lono::Cfn::Util
 
   def initialize(stack_name, options={})
-    @stack_name = stack_name
+    @stack_name = switch_current(stack_name)
     @options = options
   end
 
@@ -21,5 +21,19 @@ class Lono::Cfn::Delete
         puts "#{@stack_name.inspect} stack does not exist".colorize(:red)
       end
     end
+
+    return unless @options[:wait]
+    start_time = Time.now
+    status.wait
+    took = Time.now - start_time
+    puts "Time took for stack deletion: #{status.pretty_time(took).green}."
+  end
+
+  def status
+    @status ||= Lono::Cfn::Status.new(@stack_name)
+  end
+
+  def switch_current(stack_name)
+    Lono::Cfn::Current.name!(stack_name)
   end
 end