+++ lib/lono/cfn/util.rb
@@ -5,9 +5,9 @@ module Lono::Cfn::Util
     else
       message = case action
       when :update
-        "Are you sure you want to want to update the '#{stack_name}' stack with the changes? (y/N)"
+        "Are you sure you want to want to update the #{stack_name.colorize(:green)} stack with the changes? (y/N)"
       when :delete
-        "Are you sure you want to want to delete the '#{stack_name}' stack? (y/N)"
+        "Are you sure you want to want to delete the #{stack_name.colorize(:green)} stack? (y/N)"
       end
       puts message
       sure = $stdin.gets