+++ lib/lono/template/upload.rb
@@ -9,13 +9,14 @@ class Lono::Template::Upload
   def initialize(options={})
     @options = options
     @checksums = {}
-    @prefix = "#{folder_key}/#{Lono.env}/templates" # s3://s3-bucket/folder/developemnt/templates
+    @prefix = "#{folder_key}/#{Lono.env}/templates" # s3://s3-bucket/folder/development/templates
   end
 
   def run
     ensure_s3_setup!
     load_checksums!
 
+    say "Uploading CloudFormation templates..."
     paths = Dir.glob("#{Lono.config.output_path}/templates/**/*")
     paths.select { |p| File.file?(p) }.each do |path|
       upload(path)
@@ -95,6 +96,17 @@ class Lono::Template::Upload
   def s3_https_url(template_path)
     ensure_s3_setup!
     "https://s3.amazonaws.com/#{s3_bucket}/#{@prefix}/#{template_path}"
+  end
+
+  # used for cfn/base.rb def set_template_body!(params)
+  def s3_presigned_url(template_output_path)
+    template_path = template_output_path.sub('output/templates/','')
+    key = "#{@prefix}/#{template_path}"
+    s3_presigner.presigned_url(:get_object, bucket: s3_bucket, key: key)
+  end
+
+  def s3_presigner
+    @signer ||= Aws::S3::Presigner.new
   end
 
   # Parse the s3_folder setting and remove the folder portion to leave the